backend/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0